+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Show message when upgrading from a version with migration issues
+* Fix issue that caused wrong resting playlist sound to be set if you changed the playlist but not the track
 
 ## [4.4.1] - 2025-04-05